Refactor: Update import paths for hero, project, and skill sections in home page
