+++ src/components/module/home/project-section.tsx
@@ -1,7 +1,8 @@
 import Link from 'next/link'
 import { SiNextdotjs, SiReact, SiShadcnui, SiTailwindcss } from 'react-icons/si'
 
-import { DirectionAwareHover } from '../ui/direction-aware-hover'
+import { DirectionAwareHover } from '../../ui/direction-aware-hover'
+import TitleSection from './title-section'
 
 export default function ProjectSection() {
   const projects = [
@@ -44,13 +45,7 @@ export default function ProjectSection() {
   ]
   return (
     <div>
-      <div className="group flex rotate-6 flex-col items-center justify-center">
-        <h2 className="text-3xl font-bold transition-all group-hover:text-green-500">
-          Projects 🎨
-        </h2>
-        <div className="h-2 w-40 rounded-full bg-green-500"></div>
-        <div className="h-2 w-40 translate-x-2 rounded-full bg-indigo-500"></div>
-      </div>
+      <TitleSection title="Projects 🚀" className="rotate-6" />
       <div className="grid grid-cols-1 gap-5 pt-20 sm:grid-cols-2">
         {projects.map((item, index) => {
           return (
