+++ src/components/module/home/skill-section.tsx
@@ -9,7 +9,8 @@ import {
   SiTypescript,
 } from 'react-icons/si'
 
-import { HoverEffect } from '../ui/card-hover-effect'
+import { HoverEffect } from '../../ui/card-hover-effect'
+import TitleSection from './title-section'
 
 export default function SkillSection() {
   const skills = [
@@ -40,13 +41,7 @@ export default function SkillSection() {
   ]
   return (
     <div className="mt-20">
-      <div className="group flex -rotate-6 flex-col items-center justify-center">
-        <h2 className="text-3xl font-bold transition-all group-hover:text-green-500">
-          Skills 🔪
-        </h2>
-        <div className="h-2 w-40 rounded-full bg-green-500"></div>
-        <div className="h-2 w-40 translate-x-2 rounded-full bg-indigo-500"></div>
-      </div>
+      <TitleSection title="Skills 🔪" className="-rotate-6" />
       <HoverEffect items={skills} />
     </div>
   )
